Amendment to sunspotzooniverse_pair2images.csv
Removed the line `#Image index to pair index mapping` as well as the orphaned `#` under that line to make it slightly more painless to read this file programatically. Also I think the title might be enough to explain what this file does (pair2images) hence my suggestion to remove the in-file description.